+++ lamp/Genie/Genie/IO/OTSocket.hh
@@ -48,9 +48,9 @@ namespace Genie
 			
 			unsigned int SysPoll() const;
 			
-			int SysRead( char* data, std::size_t byteCount );
+			ssize_t SysRead( char* data, std::size_t byteCount );
 			
-			int SysWrite( const char* data, std::size_t byteCount );
+			ssize_t SysWrite( const char* data, std::size_t byteCount );
 			
 			//void IOCtl( unsigned long request, int* argp );
